Add pre-commit for linting and automatic code correction
The pre-commit utility is a git commit hook utility for taking action
on commited changes. For example, one of it's primarily uses is for
linting changes.

In this case, I've configured it with two hooks: end-of-file-fixer &
flake8. When pre-commit has been installed by the developer, these two
utilities will run when on any changed files within the commit. It
will also block the commit if their are errors.

In some cases hooks have been created to fix errors and proceed with
the commit. In others it requires user verification before
proceeding.

This commit is being created so that we are able to automate the
stylistic review of code. Thus, giving the reviewer ample freedom to
look beyond style when reviewing changes.