Temporarily install flann v2.0.0 in Dockerfile
We usually install the "next" branch for all of our github repos.  But
for flann, the unix_build.sh on the next branch isn't working after
recent changes, so use the latest tag v2.0.0 instead.

The latest version of flann requires a cmake version 3.13.0 or higher
but ubuntu 18.04 only has 3.10.2.  We will install a newer cmake version
in the future.