Temporarily install hesaff v2.0.0 in Dockerfile
We usually install the "next" branch for all of our github repos.  But
for hesaff, the unix_build.sh on the next branch isn't working after
recent changes, so use the latest tag v2.0.0 instead.

The hesaff "next" branch requires cmake version 3.13.0 or higher but ubuntu
18.04 only has version 3.10.2.  We are going to install a newer version of
cmake in the future to fix this.