Add W503 line break before operator to ignored flake8 errors
The reason for adding this to the ignored list is because W503 (Line break
occurred before a binary operator) and W504 (Line break occurred after a binary
operator) are both enabled in flake8 by default for some reason.  But they are
mutually exclusive so one has to be disabled / ignored.

According to pep8 at the time of the commit
https://www.python.org/dev/peps/pep-0008/#should-a-line-break-before-or-after-a-binary-operator

```
income = (gross_wages +
          taxable_interest +
          (dividends - qualified_dividends) -
          ira_deduction -
          student_loan_interest)
income = (gross_wages
          + taxable_interest
          + (dividends - qualified_dividends)
          - ira_deduction
          - student_loan_interest)
```

W503 (Line break occurred before a binary operator) is the correct behavior.