Use pip to install latest cmake in dockerfiles
The cmake apt package in ubuntu 18.04 is 3.10, which is too old for
flann, hesaff and vtool.  Instead of installing the cmake apt package,
we can just install it using pip with the latest version.

Remove apt install for cmake in base dockerfile and pip install cmake
and ninja in provision dockerfile.