Use the develop branch of wildbook-ia in dockerfile
I regularly build the wbia-provision with outdated wildbook-ia by
mistake because it's using the `master` branch, while the latest changes
are in the `develop` branch.  So I think it's better to use the
`develop` branch in the dockerfile.  For production, it's probably best
to use tags or what's released on pypi.