Install xvfb for running gui tests in docker files
When running doctests in the docker container:

```
DOCTEST STDOUT/STDERR
[guitool] Init new QApplication
QStandardPaths: XDG_RUNTIME_DIR not set, defaulting to '/tmp/runtime-root'
qt.qpa.screen: QXcbConnection: Could not connect to display
Could not connect to any X display.
```

On a linux host, it's possible to do this:

```
docker run --rm -ti -e DISPLAY=$DISPLAY -v /tmp/.X11-unix:/tmp/.X11-unix wildme/wbia-provision:latest /bin/bash
```

in order to run the gui doctests.  It opens new windows on the host.

If we don't care about the popups or if we're not using a linux host, we can
install and configure xvfb so it's possible to run the gui tests inside the
container.