Delete dev.py broken tune_flann doctest
The doc test looks like this:

```
>>> from wbia._devscript import *  # NOQA
>>> # build test data
>>> # execute function
>>> result = func_wrapper()
>>> # verify results
>>> print(result)
```

`wbia._devscript` doesn't have an importable `func_wrapper` function, all of
the `func_wrapper` functions are inside top level functions.  I went back to
the commit (213fede578) when this doctest was added and still couldn't find how
this test could possibly work.  On top of that, this doc test doesn't call
`tune_flann`.  I think it's ok to remove this.