Build each package individually
Creating many layers isn't going to hurt in this case. It's not the
number of layers that is a problem, but the amount of stuff in a
layer. As such, this does not create any additional stuff, but just
more layers. This then gives the developer the ability to fail and
continue from that point rather than fail and rebuild everything
before it.