Removing assert_modules test
This test is not useful. The code is a bit overengineered and
overlaps with tested and proven methods of asserting you have the
correct versions of things (i.e. pip and pkg'ing tools).

These could be simple unittests or pytest discoverable, but
instead there are several layers to peal off to see what's actually
happening here.