Add optional context manager for opendb_bg_web
We use `opendb_bg_web` in tests a lot.  We start the server then do
stuff and terminate the server at the end.  If anything fails in the
middle, the server isn't terminated.  It's better to clean that up so
the servers aren't running in the background even if the tests fail.

Add `managed=True` when calling `opendb_bg_web` to terminate the server
when the block is completed.

```
with opendb_bg_web(managed=True) as web_ibs:
    web_ibs.send_wbia_request(...)
```