Change opendb_bg_web send_wbia_request to accept non json responses
`opendb_bg_web` returns an object with a method `send_wbia_request` that
expects the responses to always be json but we have APIs that return
binary data, like images.  Add an optional argument `json=True` to
`send_wbia_request` which can be set to `False` if the response isn't
json.