Lift dumping procedure and simplify
This lifts the dumping procedure out of the class of generalized
use. It also greatly simplifies it (and from what I can tell, fixes
the `include_metadata` behavior).

I've gone with the dump/load serialization convention (e.g. used in
`json` and other apis).

I've choose to drop the argument flag in favor of including the
metadata. This is because including the metadata isn't going to hurt
anything and it's pretty core to this project's way of doing things
anyhow.