Remove __main__ check where xdoctest invocation is present
The testing path is to invoke `pytest` to run tests. If a script is
actually intended to be run as a script, it will have a `__main__`
check. Otherwise, we standardize on test invocation being done through
`pytest`.

For example, the equivalent invocation for `wbia/algo/graph/demo.py` will
be `pytest wbia/algo/graph/demo.py`.

This change removes the `__main__` check where xdoctest is the invocation.

This is in an effort to remove boilerplate redundancy code from the
files. This also helps to suggest which files are really script-like
files.

It gets one step closer to opening up CLI interface space for tooling
that is not testing related.