Replace import indirection
Just pulling the imports up. I'm being explicit about what is
imported.

Since this is __main__, it should not have any weird issues with
importing code.

Note, I've commented out the `main` function because it's not used and
name conflicts with the `main` function that is used. However, I do
see it as useful once the GUI portions of code are removed; so I'll
leave it there for the time being.