Move pytest's conftest module up to increase scope
The pytest `conftest.py` file is where pytest scoped fixtures are
registered.

The scope of `conftest.py` is the package it resides within.
Placing this in a `tests` directory on most projects, as the
convention goes, would work, but not for the wildbook-ia tests.
This is due to the tests not being scoped to a specific portion of the
code. Because of this we need to move our testing fixtures code into
the main code.

This fixes the `set_up_db` fixture to run prior to running the xdoctests.