Revise the testing workflow to work with the current setup
This provides us with a quick but thorough workflow for running the
tests against any pushed changes.

We aren't testing against all versions of python, but instead only
against our deployment version. This is intentional, as our deployment
is the critical focus for now. Later adjustments can be made to do a
full matrix build if necessary. More than likely matrix testing would
be on pull request creation. This tries to simply get us to a place
where the tests are routinely running.