Migrate away from file based sql connections
This opens up the controller class to instantiation from a URI that
could be specifying the database run in memory, from file or via a
service.

Note, this new `from_uri` class method works around `__init__`, which
is locked into being file focused. Rather than dance around making
`__init__` compatible, I've chosen to simplify the procedure, because
some of the other things happening in there will need to be refocused
in order for this object to work with an external service.