Rename main_module to entry_points
This is to clarify that this module is for entry-points.

My assumption about a module to be used externally would be that it
not be imported internally. However, that's not exactly the case
here. Therefore, I'm moving in the direction that will make the
assumption correct and make the code more understandable and intuitive
in the process.