Replace implicit parsing of cli args with explicity invocation
With this change, these functions no longer require the wbia.params module
to have already populated the 'args' variable. Instead, we ask each user
of the cli arguments to explicitly parse the arguments.

This effectively means that a followup commit can remove the population
logic from wbia.params in a later commit.