Remove import loading behavior of wbia.params module
This change removes the on import population of the 'args' (and 'unknown')
variable. This means users of the arguments will need to explicitly
call 'parse_args' in order to use the cli argument values.
Note, a previous commit makes this change.

Removal of the implicit behavior should :crossed_fingers: open up
the CLI space. In other words, these changes should allow other CLI
tools to be added to this project without wbia.params encroaching
and bullyingly using the interface space.