Clean up the method that makes create table sql
I was trying to make sense of some of the things happening in this
method. Best way for me to do that was to clean it up.

The method now has required arguments, which were basically required
arguments given all the checking that was happening in the method
itself. This also make them typed arguments to remove the additional
type checking in method. I've kept the value checking, but made this
ValueErrors rather than AssertionErrors, which is typically what they
should be anyway.

The method AFAICT behaves the exact same way as before, it's just
easier to read and anticipate what it is doing.