Add docker-compose.yml with postgres and runtime image
`docker-compose.yml` has `postgres:10` and the runtime image.  It's
possible to use `wildme/wbia-provision:latest` instead, see commented
out line.

`docker-compose.env` has the environment variables for setting up the
postgres database.

`init-db.sh` is a bash file to create the database and the user.  It is
mounted in the postgres container and is automatically run.