Deprecate tablenames methods in favor of table_names property
This is more of a naming change than anything, but it also helps to
guarantees the use of `_tablenames` is no longer used outside
of this class.

Invalidation of the cached table names is now done implicitly through
the use of `_is_table_names_stale`. This allows known manipulaters of
the table names to invalidate the cache rather than external api users
of the tables names data to invalidate it on invocation.