Add wbia-cnn to runtime requirements
`wbia-cnn` adds `generate_species_background_mask` to `IBEISController`
which seems to be central to a lot of the functionality in wildbook-ia.

We have 13 tests that don't work without `wbia-cnn`.  It seems that
`wbia-cnn` isn't like other plugins and should be installed as a runtime
requirement.

`wbia-cnn` requires `Lasagne` but the version on pypi is too old so we
have embedded it in `wbia-cnn`.