Execute runtime command as wbia
This calls to `gosu`, an execution setter, that will transition to
executing the command as the `wbia` user even though the container now
starts and runs as root. This entrypoint script will ensure the
process is running as the user rather than root.

I could have also used `su-exec`, but it's not already packaged in
debian.