Revise and test delete & delete_rowids methods
Revised to use sqlalchemy. Also delete_rowids is now a submethod of
delete.

Sidenote, this is a terribly inefficient way to do a delete, one at a
time. But I can't seem to get the in operator with an array to
work. For example, `where id in (1, 2, 3)` doesn't work because
python's sqlite3 lib fails to translate a list to an array. And
sqlalchemy won't push it through either. So we're stuck doing what it
was doing before.