Rewrite the sql controller's executemany method
This has been rewritten to simplify things. This gets us one
step forward by making the logic sqlalchemy compatible. We're
sacrificing the logging bits for simplicity.

I'm looking to deprecate this method in a later commit. This will be
kept around to provide backwards compatiblity with the current users
of it. The new method will be a generator, which will give us the
ability to bring back the logging facilities if necessary.