Update _make_add_table_sqlstr for sqlalchemy
Updating some of the internals to be consistent and readable at a
glance. Also, documenting the code blocks.

I'm removing the docstr test because it doesn't test for anything, but
maybe that the method output something. So one less test to have
hanging around. Besides the previous commit submits a unittest for
this method.