Test and correct the sql controller's get method
Note, the method's doctest is still failing, because it depends on
other moving parts that aren't fixed yet. I'm tempted to drop the
doctest because it's not as comprehensive as the unittests I've added
in this changeset, but I'll leave it for now.

The changes in the method were made to make the method compatible with
sqlalchemy.

IMO the `assume_unique` bit of this function should be an entirely
new method, but if I understand correctly, it may have been put in as
an optimization.