Test executeone's no results expectation
As noted in the test's comments,
this breaks backwards compatiblity,
where no results returned an empty list (i.e. `[]`).
IMO returning None is correct,
because that's the expectation from `fetchone`'s DBAPI spec.

That is I see `executeone` as the shortcut for calling `execute` and
`fetchone`, which I believe is a fairly correct assumption. Therefore,
the returned value expectations should match that of
`fetchone` (ignoring the weird `unpack_scalars` behavior).