Fix problematic interpretation of executeone's behavior
My interpretation of the `executeone` method's behavior has been
slightly incorrect. I believed we were shortcutting the
procedure to `execute` and `fetchone`. This turns out to not exactly
be true. I've corrected this problem by retaining the, IMO wrong,
results of an empty list. But I've added the ability through
`use_fetchone_behavior` to correspond with expectations defined in the
DBAPI v2 spec.