Build the update statment using with typed columns
This updates/rewrites the set method's SQL to build from the Table
instance. This solves the issue where the id is something other than
a simple integer. For example, this was erroring when the id was a
UUID object. With this change in place, the id parameter gets the type
treatment of the column it is trying to compare with.