Fix sql controller's set method to accept mistyped input
The problem is that sometimes the method is given inputs that are
sequence of sequences, while other times it gets a sequence of item
values. It's a terribly inconsistent usage from many places.

This puts in a backwards compatiblity set, while at the same time
letting the code that follows read as if the problem didn't exist.

The other weird thing is that this method sometimes gets called with
an empty list of values or ids. This is why the conditions here check
the variable for _somethingness_ before looking inside it.