Correct the output of get_column
This is a straight text query without knowledge of how to cast the results to python types.
This change uses the table definition to ensure the correct python typed value is returned.

An example of this failing would be around a selection of UUID values.
Because we store them as bytes, they don't even come across as inferable.