Cast numpy.ndarray integer values to int
Upon the SQLController's `get` method being given a `numpy.ndarray` or
`list` of `numpy.int*` values, cast all the values to `int`.
This is only needed for `rowid`, because it's not a real
column. Otherwise the sqlalchemy types would convert the values to the
correct sql type.

This resolves about 70+ failing tests.