Collapse _CoreDependencyCache into DependencyCache
There was no benefit to these classes being separate. The lack of
constructor on `_CoreDependencyCache` with the dependence on attributes
defined by `DependencyCache` was bothering me.

Rather than dance around the issue, I'm just making them one
class. I'm not sure it'd be easy to subclass `_CoreDependencyCache` in
that state anyhow.

Because the new `as_named` constructor depends on the controller, I've
added a `DummyController` to all the example_depcache* cases. This
implements enough of the controller to instantiate a depcache. My hope
is that this dependence won't last, but is a bridge to a different
type of solution moving forward.