Remove built documentation
I'm removing the built documentation in favor of building it with the
latest code. This will routinely become out-of-sync, as it is at the
moment. The better solution is to build it as a CI job and provide
clear instructions to users of this code on how to built it from
scratch. This ensures that the latest documentation is reflective of
the latest code.