Update postgres sequence owner when modifying tables
The way we modify tables is by creating a new table, dropping the original
table and then renaming the new table to the original table.  This turns out to
be more complicated in postgresql.  We have lots of "self-increment" ("serial"
in postgresql) fields, e.g. rowid.  In postgresql, these are done by having a
"sequence", e.g. `images_rowid_seq`.  When we create a new table and use the
existing sequence and then try to drop the old table with "cascade", the
sequence is dropped and the `images_rowid` field is not "self-incrementing"
anymore.  If we don't drop the old table with "cascade":

```
testdb1=# drop table images;
ERROR:  cannot drop table images because other objects depend on it
DETAIL:  default for table images_temp2603a7eb column rowid depends on sequence images_rowid_seq
HINT:  Use DROP ... CASCADE to drop the dependent objects too.
```

The way to fix this is by changing the sequences to be owned by the new
table.  That way the old table can be dropped and the
"self-incrementing" property is retained in the new table.