Fix the containers entrypoint
Reference the entry-point script but not the default command. Instead
put the default command as the CMD value. This allows the user to
override the command without overriding the entrypoint.

Think of the entrypoint as a decorator function. In some cases we want
to run a slightly different command, but we still want the same
general behavior.