Refactor primary species lookup to a single query
There are two things wrong with how this method is written: 1) does a
shit load of database queries 2) orders and aggregates outside the database.

This does 1 + (N*1) + (N*1) queries against the database, where N is
the number of annotations (aids). So for the manta slim dataset this
is about 97k queries for this one
operation (i.e. get_primary_species). It's kinda surprising it did so
well.

The ordering and aggregation was done outside of the database. This
effectively ignores one of the great powers of sql databases.

This change reduces the operation to one query, letting the database
power through joining the data, aggregating it and ordering it.