Remove _init_config
This method call essentially initializes the default database
species.

The only other thing it did was to re-init/reset the
tablecache. That is also called in `_initialize_self`.

Tests pass, so no need to keep it hanging around. For now I'm leaving
the `_init_config` definition. It'll get vaccumed in a cleanup phase.