Add the --db-uri commandline option and move backup logic
The `--db-uri` option is added to allow for a postgresql connection
URI to be given in order for the code to use the postgresql database.

This maintains the concept of "db dir" in the code, because it's still
required for cache storage and such. This concept is more precisely
turning into a `/var/` space. But we'll need to maintain the legacy
term because it's everywhere.

This pulls the the daily backup on startup logic up closer to the user's
requested operation. A followup change might break this out into a
separate script that can be run in a cronjob.