Remove from_uri in favor of init on sqlcontroller to name it
Adds a ``name`` attribute to the SQLController in order to make the
sql controller name aware. When using sqlite this attribute is
benign. The attribute is used as the schema name when using Postgres.

I'm using the database filename as the controller name. It happens to
be available and is the canonical identifier.

The controller can now use the given uri to connect directly to
postgres without deriving the connection uri from a sqlite based uri.
This completes the removal of support for the POSTGRES environment
variable. The code is now capable of running the same from production
to test.