Fix REAL values losing precision after migration to postgres
After migrating the sqlite database to postgresql, the REAL type is
migrated to postgresql's REAL type but postgresql's REAL type only
stores 6 digits causing values to lose precision for example
`-1.4120325749880058` becomes `-1.41203`.  We need the DOUBLE PRECISION
type instead which stores 15 digits.

pgloader has a way to add a transform rule but there seems to be a bug:

```
FATAL error: At

    CAST type uuid to uuid using wbia-uuid-bytes-to-uuid,
         type real to double precision using float-to
                             ^ (Line 14, Column 27, Position 397)
```

So instead we are changing the REAL columns in sqlite to DOUBLE because
pgloader has a built in transform rule to migrate DOUBLE to DOUBLE
PRECISION in postgresql.

The values are still not exact, `3.9269908169872414` becomes
`3.92699081698724`.