Adjust run_pgloader to be state unaware
Removing the temporary directory as input from the function so that
it's more standalone. This makes it so that the function can be can
without knowing anything about the world around it.
In theory it could now be given a remote sqlite database. Not sure if
that's of any benefit, but ya never know.