Comment out localhost from /etc/hosts in docker container
We are using localhost as the nginx service name so we can avoid using
http://houston:5000/ in development.

From houston we call the detection API
`/api/engine/detect/cnn/lightnet/` with a list of image urls and a
callback url.  In development, "http://localhost/" is the url for the
site, but acm wouldn't be able to access a url at localhost unless we
remove the definition of localhost from `/etc/hosts`.